VR: Optimize Header and update CLI interface (breaking)
This is a breaking change for client and cluster init that:

* adds `Header.head` as a u128,
* adds `Header.version` as a u8,
* resizes `Header.cluster` from u128 to u32,
* accepts `Header.cluster` now being zero (no longer a UUID),
* resizes `Header.view` from u64 to u32,
* resizes `Header.replica` from u16 to u8, and
* adds a client table fix.

The motivation behind this is to:

* add space in the header to support end-to-end integrity
  checking in the client to match replies back to requests,
  and prevent future issues with protocol versioning, and

* make the cluster CLI more memorable and easier to use:

  --cluster=0 --replica=0 --addresses=3001,3002,3003

+++ src/vr/journal.zig
@@ -17,12 +17,12 @@ pub const Journal = struct {
     pub const Read = struct {
         replica: *Replica,
         completion: Storage.Read,
-        callback: fn (self: *Replica, prepare: ?*Message, destination_replica: ?u16) void,
+        callback: fn (self: *Replica, prepare: ?*Message, destination_replica: ?u8) void,
 
         message: *Message,
         op: u64,
         checksum: u128,
-        destination_replica: ?u16,
+        destination_replica: ?u8,
     };
 
     pub const Write = struct {
@@ -82,7 +82,7 @@ pub const Journal = struct {
 
     allocator: *Allocator,
     storage: *Storage,
-    replica: u16,
+    replica: u8,
     size: u64,
     size_headers: u64,
     size_circular_buffer: u64,
@@ -119,7 +119,7 @@ pub const Journal = struct {
     pub fn init(
         allocator: *Allocator,
         storage: *Storage,
-        replica: u16,
+        replica: u8,
         size: u64,
         headers_count: u32,
     ) !Journal {
@@ -427,10 +427,10 @@ pub const Journal = struct {
     pub fn read_prepare(
         self: *Journal,
         replica: *Replica,
-        callback: fn (replica: *Replica, prepare: ?*Message, destination_replica: ?u16) void,
+        callback: fn (replica: *Replica, prepare: ?*Message, destination_replica: ?u8) void,
         op: u64,
         checksum: u128,
-        destination_replica: ?u16,
+        destination_replica: ?u8,
     ) void {
         if (op > replica.op) {
             self.read_prepare_notice(op, checksum, "beyond replica.op");
